+++ src/game/anim/Anim_Blend.cpp
@@ -899,7 +899,9 @@ void idAnim::CallFrameCommands( idEntity *ent, int from, int to ) const {
 // abahr:
 				case FC_EVENTFUNCTION_ARGS: {
 					assert( command.event );
-					ent->ProcessEvent( command.event, (int)command.parmList );
+// jmarshall - 64bit
+					ent->ProcessEvent( command.event, (intptr_t)command.parmList );
+// jmarshall end
 					break;
 				}
 // bdube: support indirection and simplify
